+++ tests/testthat/test-vroom_write.R
@@ -1,5 +1,19 @@
 context("test-vroom_write.R")
 
+test_that("empty inputs just return themselves without writing anything", {
+  out <- tempfile()
+
+  no_rows <- mtcars[FALSE, ]
+  no_cols <- mtcars[, FALSE]
+  no_rows_or_cols <- mtcars[FALSE, FALSE]
+
+  expect_equal(vroom_write(no_rows, out), no_rows)
+  expect_equal(vroom_write(no_cols, out), no_cols)
+  expect_equal(vroom_write(no_rows_or_cols, out), no_rows_or_cols)
+
+  expect_false(file.exists(out))
+})
+
 test_that("strings are only quoted if needed", {
   x <- c("a", ',')
 
@@ -156,7 +170,7 @@ test_that("Can change the escape behavior for quotes", {
 })
 
 test_that("hms NAs are written without padding (#930)", {
-  df <- data.frame(x = hms::as.hms(c(NA, 34.234)))
+  df <- data.frame(x = hms::as_hms(c(NA, 34.234)))
   expect_equal(vroom_format(df), "x\nNA\n00:00:34.234\n")
 })